chore(example): code clean up
