+++ example/lib/widgets/demo_1.dart
@@ -1,0 +1,64 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_offline/flutter_offline.dart';
+
+class Demo1 extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return OfflineBuilder(
+      connectivityBuilder: (
+        BuildContext context,
+        ConnectivityResult connectivity,
+        Widget child,
+      ) {
+        final bool connected = connectivity != ConnectivityResult.none;
+        return Stack(
+          fit: StackFit.expand,
+          children: [
+            child,
+            Positioned(
+              height: 32.0,
+              left: 0.0,
+              right: 0.0,
+              child: AnimatedContainer(
+                duration: const Duration(milliseconds: 350),
+                color: connected ? Color(0xFF00EE44) : Color(0xFFEE4400),
+                child: AnimatedSwitcher(
+                  duration: const Duration(milliseconds: 350),
+                  child: connected
+                      ? Text('ONLINE')
+                      : Row(
+                          mainAxisAlignment: MainAxisAlignment.center,
+                          children: <Widget>[
+                            Text('OFFLINE'),
+                            SizedBox(width: 8.0),
+                            SizedBox(
+                              width: 12.0,
+                              height: 12.0,
+                              child: CircularProgressIndicator(
+                                strokeWidth: 2.0,
+                                valueColor:
+                                    AlwaysStoppedAnimation<Color>(Colors.white),
+                              ),
+                            ),
+                          ],
+                        ),
+                ),
+              ),
+            ),
+          ],
+        );
+      },
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: <Widget>[
+          new Text(
+            'There are no bottons to push :)',
+          ),
+          new Text(
+            'Just turn off your internet.',
+          ),
+        ],
+      ),
+    );
+  }
+}
