+++ example/lib/widgets/demo_2.dart
@@ -1,0 +1,44 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_offline/flutter_offline.dart';
+
+class Demo2 extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return OfflineBuilder(
+      connectivityBuilder: (
+        BuildContext context,
+        ConnectivityResult connectivity,
+        Widget child,
+      ) {
+        if (connectivity == ConnectivityResult.none) {
+          return Container(
+            color: Colors.white,
+            child: Center(
+              child: Text(
+                "Oops, \n\nNow we are Offline!",
+                style: TextStyle(color: Colors.black),
+              ),
+            ),
+          );
+        } else {
+          return child;
+        }
+      },
+      builder: (BuildContext context) {
+        return Center(
+          child: new Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: <Widget>[
+              new Text(
+                'There are no bottons to push :)',
+              ),
+              new Text(
+                'Just turn off your internet.',
+              ),
+            ],
+          ),
+        );
+      },
+    );
+  }
+}
